feat: Move snooze + FCM model types to domain module (Phase 19)
Move SnoozeDurationUIOption, SnoozeDurationServerOption, toServer(),
and String.toFcmTopic() from androidApp to domain/src/commonMain/.
These are pure Kotlin with zero Android dependencies.
