+++ AndroidGarage/androidApp/src/test/java/com/chriscartland/garage/fcm/FcmTopicTest.kt
@@ -17,6 +17,7 @@
 
 package com.chriscartland.garage.fcm
 
+import com.chriscartland.garage.domain.model.toFcmTopic
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertTrue
 import org.junit.Test
