+++ AndroidGarage/domain/src/commonMain/kotlin/com/chriscartland/garage/domain/model/SnoozeNotificationsModel.kt
@@ -15,7 +15,7 @@
  *
  */
 
-package com.chriscartland.garage.snoozenotifications
+package com.chriscartland.garage.domain.model
 
 import kotlin.time.Duration
 import kotlin.time.Duration.Companion.hours
